#2 Fix concurrency issue
Previously you may face error messages like 'This continuation has already been invoked [...] line 143' or 'Use of uninitialized value of type Mu in string context. [...] line 112'. It appears because of laziness of the sequence operator for $digits list in the Pattern package.

Now the sequence operator is changed to explicit list initialization.